requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
ccxt==4.2.4
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-debug-toolbar==4.2.0
django-environ==0.9.0
django-extensions==3.2.3
django-filter==21.1
django-import-export==3.0.2
django-redis==5.2.0
django-rest-framework==0.1.0
django-silk==5.0.4
djangorestframework==3.14.0
drf-yasg==1.21.4
ecdsa==0.18.0
et-xmlfile==1.1.0
frozenlist==1.3.3
gevent==22.10.2
gprof2dot==2022.7.29
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
line-profiler==4.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
minio==7.1.12
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.36
psycopg==3.1.12
pyasn1==0.4.8
pycares==4.3.0
pycodestyle==2.11.1
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
snakeviz==2.2.0
sqlparse==0.4.3
tablib==3.4.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
django-cors-headers==4.3.0
pybit==5.6.2
django-admin-sortable2==2.1.10